pyproject.toml
[project]
name = "lingua"
version = "0.1.0"
description = "an AI powered app for learning languages with AI"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "chromadb>=1.4.0",
    "google-adk>=1.22.0",
    "google-generativeai>=0.8.6",
    "gtts>=2.5.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
